feat(utils.ts): вынести функции calculateAge и formatDateShort в файл src/lib/utils.ts для переиспользования в других компонентах
