+++ src/pages/Result/components/cardSection/CardSection.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import styles from './CardSection.module.css';
+import { calculateAge, formatDateShort } from '../../../../lib/utils';
 
 export type TCardSection = {
   className?: string;
@@ -29,34 +30,6 @@ export type TCardSection = {
 //   whereLives: 'В семье',
 //   socialFeatures: 'полная семья, есть старшая сестра',
 // };
-
-const calculateAge = (dateString: string) => {
-  const today = new Date();
-  const date = new Date(dateString);
-  const age = today.getFullYear() - date.getFullYear();
-  let yearsText = 'лет';
-  let count = age % 100;
-  if (count >= 5 && count <= 20) {
-    yearsText = 'лет';
-  } else {
-    count = count % 10;
-    if (count === 1) {
-      yearsText = 'год';
-    } else if (count >= 2 && count <= 4) {
-      yearsText = 'года';
-    }
-  }
-  const personAge: string = String(age) + ' ' + yearsText;
-  return personAge;
-};
-
-const formatDateShort = (dateString: string): string => {
-  return new Date(dateString).toLocaleDateString('ru-RU', {
-    day: '2-digit',
-    month: '2-digit',
-    year: 'numeric',
-  });
-};
 
 export const CardSection: React.FC<TCardSection> = ({
   className,
